feat: Admin, Driver, User 관련 파일들 수정 -> 1.Dto는 최소한의 요소들로만 구현했음, ex) 아이디, 이름, 비밀번호, Role 등등 필수적인 부분만. 2.비밀번호 다시 해싱해서 사용하는 버전으로 구현했는데 이부분은 안되면 다시 소통이 필요할거같다.  3.Service 전체적으로 바뀐 요소들 맞춰서 수정 및 디벨롭. ex) id는 Long형으로 전부 통일. Role이 List값에서 Enum형식으로 바뀌며 Role 관련 함수 수정 및 요소들 간소화로 인한 생성자 수정 등등.  4.Role Enum 형식으로 바꿨고 Entity파일에 Role 추가. 5.Operator와의 관계성 일단 보류. -> 간소화 하기 위해 일단 쳐냈음.

+++ src/main/java/com/drive/backend/drive_api/repository/OperatorRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface OperatorRepository extends JpaRepository<Operator, Integer> {
+public interface OperatorRepository extends JpaRepository<Operator, Long> {
     Optional<Operator> findByOperatorCode(String operatorCode);
 }
